Added the ability to load more listings
I added state for the visible jobs and currently visible page numbers. If the user
wants to see more listings, they can hit the load more button. The load more button
will be removed from the screen once the user has added all of the listings to the
grid. Also, I moved the load more button into the App component so it can be closer
to the state it relies on, and I made slight styling changes to accomodate this
change.