Created applied filter's view
I created a new state object that will hold onto the currently applied filters
so the user always knows what filters are applied to the job grid. I changed the
structure of how jobs are filtered and passed in a filter dispatch function to
the forms. Also, I added new icons for the filter tabs.